refactor: dry parse immediate value
